Move `specialist_document` to `section`
