+++ app/models/document_factory_registry.rb
@@ -6,7 +6,7 @@ require "validators/null_validator"
 require "builders/manual_document_builder"
 require "manual_with_documents"
 require "slug_generator"
-require "specialist_document"
+require "section"
 
 class DocumentFactoryRegistry
   def self.validatable_document_factories
