+++ app/models/section.rb
