+++ spec/models/section_spec.rb
